superpatch-backend/requirements.txt
fastapi
uvicorn[standard]
langgraph
langgraph-checkpoint-postgres
langchain-google-genai
langchain-core
langchain-text-splitters
psycopg[binary]
supabase
python-dotenv
pyyaml
httpx
pytest
pytest-asyncio
